requirements.txt
django==5.0.6
python-dotenv==1.0.1
requests==2.32.3
osu.py==4.0.2
whitenoise[brotli]==6.7.0
PyNaCl==1.5.0
psycopg[binary]==3.1.18
black==25.1.0
